refactor: move create_general_modelconfig out of the ensemble utils
Resolves the [TODO] carried in gen_qa_models, gen_qa_implicit_models and
gen_qa_sequence_models. Five generators build a plain model config with this
function and only one of them has anything to do with ensembles, so reaching
for it through `emu` named the file it happened to live in rather than what it
does.

Moved to gen_common with dtype_str, reshape_str and label_str, which only it
uses -- leaving those behind would have made gen_common import emu while emu
imports gen_common. gen_ensemble_model_utils imports the two it still needs.

gen_qa_implicit_models no longer references emu at all, so its import goes too.

Behaviour-preserving: the ONNX and OpenVINO stages regenerate 751 models,
including 598 ensembles, whose config.pbtxt are byte-identical to the tree
built before the move.
