+++ qa/common/gen_ensemble_model_utils.py
@@ -30,7 +30,7 @@ import os
 
 import numpy as np
 import test_util as tu
-from gen_common import np_to_model_dtype
+from gen_common import create_general_modelconfig, dtype_str, np_to_model_dtype
 
 BASIC_ENSEMBLE_TYPES = ["simple", "sequence", "fan"]
 
@@ -1091,128 +1091,5 @@ def create_nop_tunnel_modelconfig(models_dir, tensor_shape, tensor_dtype):
         cfile.write(config)
 
 
-def create_general_modelconfig(
-    model_name,
-    platform,
-    max_batch,
-    input_dtypes,
-    input_shapes,
-    input_model_shapes,
-    output_dtypes,
-    output_shapes,
-    output_model_shapes,
-    label_filenames,
-    backend=None,
-    version_policy=None,
-    default_model_filename=None,
-    instance_group_str="",
-    force_tensor_number_suffix=False,
-):
-    assert len(input_dtypes) == len(input_shapes)
-    assert len(input_model_shapes) == len(input_shapes)
-    assert len(output_dtypes) == len(output_shapes)
-    assert len(output_model_shapes) == len(output_shapes)
-    assert len(label_filenames) == len(output_shapes)
-
-    # Unpack version policy
-    version_policy_str = "{ latest { num_versions: 1 }}"
-    if version_policy is not None:
-        type, val = version_policy
-        if type == "latest":
-            version_policy_str = "{{ latest {{ num_versions: {} }}}}".format(val)
-        elif type == "specific":
-            version_policy_str = "{{ specific {{ versions: {} }}}}".format(val)
-        else:
-            version_policy_str = "{ all { }}"
-
-    default_model_filename_str = ""
-    if default_model_filename is not None:
-        default_model_filename_str = 'default_model_filename: "{}"'.format(
-            default_model_filename
-        )
-
-    # If backend is specified use backend instead of platform
-    if backend is not None:
-        key = "backend"
-        val = backend
-    else:
-        key = "platform"
-        val = platform
-
-    config = """
-name: "{}"
-{}: "{}"
-max_batch_size: {}
-version_policy: {}
-{}
-{}
-""".format(
-        model_name,
-        key,
-        val,
-        max_batch,
-        version_policy_str,
-        default_model_filename_str,
-        instance_group_str,
-    )
-
-    for idx in range(len(input_dtypes)):
-        idx_str = ""
-        if len(input_dtypes) != 1 or force_tensor_number_suffix:
-            idx_str = str(idx)
-        config += """
-input [
-  {{
-    name: "INPUT{}"
-    data_type: {}
-    dims: [ {} ]
-    {}
-  }}
-]""".format(
-            idx_str,
-            dtype_str(input_dtypes[idx]),
-            tu.shape_to_dims_str(input_shapes[idx]),
-            reshape_str(input_shapes[idx], input_model_shapes[idx]),
-        )
-
-    for idx in range(len(output_dtypes)):
-        idx_str = ""
-        if len(input_dtypes) != 1 or force_tensor_number_suffix:
-            idx_str = str(idx)
-        config += """
-output [
-  {{
-    name: "OUTPUT{}"
-    data_type: {}
-    dims: [ {} ]
-    {}
-    {}
-  }}
-]""".format(
-            idx_str,
-            dtype_str(output_dtypes[idx]),
-            tu.shape_to_dims_str(output_shapes[idx]),
-            reshape_str(output_shapes[idx], output_model_shapes[idx]),
-            label_str(label_filenames[idx]),
-        )
-    return config
-
-
 def repeat(obj, cnt):
     return [obj] * cnt
-
-
-def dtype_str(dtype):
-    return dtype if isinstance(dtype, str) else np_to_model_dtype(dtype)
-
-
-def reshape_str(shape, model_shape):
-    if model_shape is None or shape == model_shape:
-        return ""
-    return "reshape: {{ shape: [ {} ] }}".format(tu.shape_to_dims_str(model_shape))
-
-
-def label_str(label):
-    if label is None:
-        return ""
-    return 'label_filename: "{}"'.format(label)
